Refactor the global usage code into a separate module, fix inheriting from the callees
