+++ src/front/mod.rs
@@ -1,5 +1,7 @@
 //! Parsers which load shaders into memory.
 
+mod global_usage;
+
 #[cfg(feature = "glsl-in")]
 pub mod glsl;
 #[cfg(feature = "spv-in")]
