+++ src/proc/interface.rs
@@ -216,105 +216,3 @@ where
         }
     }
 }
-
-struct GlobalUseVisitor<'a>(&'a mut [crate::GlobalUse]);
-
-impl Visitor for GlobalUseVisitor<'_> {
-    fn visit_expr(&mut self, expr: &crate::Expression) {
-        if let crate::Expression::GlobalVariable(handle) = expr {
-            self.0[handle.index()] |= crate::GlobalUse::READ;
-        }
-    }
-
-    fn visit_lhs_expr(&mut self, expr: &crate::Expression) {
-        if let crate::Expression::GlobalVariable(handle) = expr {
-            self.0[handle.index()] |= crate::GlobalUse::WRITE;
-        }
-    }
-}
-
-impl crate::Function {
-    pub fn fill_global_use(&mut self, globals: &Arena<crate::GlobalVariable>) {
-        self.global_usage.clear();
-        self.global_usage
-            .resize(globals.len(), crate::GlobalUse::empty());
-
-        let mut io = Interface {
-            expressions: &self.expressions,
-            local_variables: &self.local_variables,
-            visitor: GlobalUseVisitor(&mut self.global_usage),
-        };
-        io.traverse(&self.body);
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::{
-        Arena, Expression, GlobalUse, GlobalVariable, Handle, Statement, StorageAccess,
-        StorageClass,
-    };
-
-    #[test]
-    fn global_use_scan() {
-        let test_global = GlobalVariable {
-            name: None,
-            class: StorageClass::Uniform,
-            binding: None,
-            ty: Handle::new(std::num::NonZeroU32::new(1).unwrap()),
-            init: None,
-            interpolation: None,
-            storage_access: StorageAccess::empty(),
-        };
-        let mut test_globals = Arena::new();
-
-        let global_1 = test_globals.append(test_global.clone());
-        let global_2 = test_globals.append(test_global.clone());
-        let global_3 = test_globals.append(test_global.clone());
-        let global_4 = test_globals.append(test_global);
-
-        let mut expressions = Arena::new();
-        let global_1_expr = expressions.append(Expression::GlobalVariable(global_1));
-        let global_2_expr = expressions.append(Expression::GlobalVariable(global_2));
-        let global_3_expr = expressions.append(Expression::GlobalVariable(global_3));
-        let global_4_expr = expressions.append(Expression::GlobalVariable(global_4));
-
-        let test_body = vec![
-            Statement::Return {
-                value: Some(global_1_expr),
-            },
-            Statement::Store {
-                pointer: global_2_expr,
-                value: global_1_expr,
-            },
-            Statement::Store {
-                pointer: expressions.append(Expression::Access {
-                    base: global_3_expr,
-                    index: global_4_expr,
-                }),
-                value: global_1_expr,
-            },
-        ];
-
-        let mut function = crate::Function {
-            name: None,
-            arguments: Vec::new(),
-            return_type: None,
-            local_variables: Arena::new(),
-            expressions,
-            global_usage: Vec::new(),
-            body: test_body,
-        };
-        function.fill_global_use(&test_globals);
-
-        assert_eq!(
-            &function.global_usage,
-            &[
-                GlobalUse::READ,
-                GlobalUse::WRITE,
-                GlobalUse::WRITE,
-                GlobalUse::READ,
-            ],
-        )
-    }
-}
